+++ src/RcppModelData.cpp
@@ -59,23 +59,22 @@ bool isRcppPtrNull(SEXP x) {
 }
 
 // [[Rcpp::export("getNumberOfStrata")]]
-size_t ccdGetNumberOfStrata(Environment x) {			
+int ccdGetNumberOfStrata(Environment x) {			
 	XPtr<bsccs::ModelData> data = parseEnvironmentForPtr(x);	
-	return data->getNumberOfPatients();
+	return static_cast<int>(data->getNumberOfPatients());
 }
 
 // [[Rcpp::export("getNumberOfCovariates")]]
-size_t ccdGetNumberOfColumns(Environment x) {	
+int ccdGetNumberOfColumns(Environment x) {	
 	XPtr<bsccs::ModelData> data = parseEnvironmentForPtr(x);	
-	return data->getNumberOfColumns();
+	return static_cast<int>(data->getNumberOfColumns());
 }
 
 // [[Rcpp::export("getNumberOfRows")]]
-size_t ccdGetNumberOfRows(Environment x) {	
+int ccdGetNumberOfRows(Environment x) {	
 	XPtr<bsccs::ModelData> data = parseEnvironmentForPtr(x);	
-	return data->getNumberOfRows();
+	return static_cast<int>(data->getNumberOfRows());
 }
-
 
 // [[Rcpp::export(".ccdSum")]]
 std::vector<double> ccdSum(Environment x, const std::vector<long>& covariateLabel) {
@@ -108,7 +107,7 @@ List ccdNewSqlData(const std::string& modelTypeName) {
 
 
 // [[Rcpp::export(".appendSqlCcdData")]]
-size_t ccdAppendSqlData(Environment x,
+int ccdAppendSqlData(Environment x,
         const std::vector<long>& oStratumId,
         const std::vector<long>& oRowId,
         const std::vector<double>& oY,
@@ -121,7 +120,7 @@ size_t ccdAppendSqlData(Environment x,
     using namespace bsccs;
     XPtr<ModelData> data = parseEnvironmentForPtr(x); // TODO This can cause a slice error when sent a ModelData
     size_t count = data->append(oStratumId, oRowId, oY, oTime, cRowId, cCovariateId, cCovariateValue);
-    return count;
+    return static_cast<int>(count);
 }
 
 // [[Rcpp::export(".ccdReadData")]]
